// gaveta/dispositivo/00_devise_app/src/app/models/projeto.ts
import { Chat } from './chat';

export class Projeto{
    _id: string;
    nome: string;
    descricao: string;
    tipoCasa: string;
    metragem: number;
    qtdComodos: number;
    fase1: string;
    fase2: string;
    fase3: string;
    fase4: string;
    chat: string;
    idCliente: string;
    comments: string;
}